requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-openai
openai
python-dotenv
chromadb
